Esc key will quit to menu.  Can't play using keyboard though.
I'd added an input controller to the level_manager to look for the esc key during the game so it could quit back to the menu.  But Unity's input_controller_manager was then triggering the add_player event, passing it the level manager instead of a player, so crashes were happening and it would then no longer let any player join.  I put in a test, so it doesn't crash anymore, but it also won't create a player for the keyboard input, as it already triggered from the level manager.  Probably worth it, but I may look into changing things so it'll work.
Also fixed up the singleton patterns on the sfx,music and gamemanager as they were leaving the game objects behind only destroying the component on the objects.